backend -> develop 연동 위한 merge 2025.09.18

See merge request s13-fintech-finance-sub1/S13P21A509!41

+++ backend/src/main/java/com/ssafy/keeping/domain/charge/dto/response/PrepaymentResponseDto.java
@@ -15,7 +15,7 @@ public class PrepaymentResponseDto {
     private String transactionUniqueNo;
     private Long storeId;
     private String storeName;
-    private BigDecimal paymentAmount;
+    private Long paymentAmount;
     private LocalDateTime transactionTime;
-    private BigDecimal remainingBalance;
+    private Long remainingBalance;
 }
